型定義を移動＆追加
