+++ blog-next/interfaces/index.ts
@@ -10,20 +10,4 @@ export type User = {
   name: string
 }
 
-export type Article = {
-  id: number
-  member: number
-  title: string
-  content: string
-  article_image: string
-  article_tags: number[]
-  is_active: boolean
-  pub_date: string
-}
 
-export type ArticleTag = {
-  id: number
-  name: string
-  content: string
-  pub_date: string
-}
